+++ graphdatascience/tests/integration/test_graph_ops.py
@@ -424,14 +424,24 @@ def test_graph_nodeProperty_stream_raise_error_with_duplicate_keys(gds: GraphDat
         gds.graph.nodeProperty.stream(G, "x", db_node_properties=["x", "z", "name"], concurrency=2)
 
 
-@pytest.mark.enterprise
+def test_graph_nodeProperty_fail_on_duplicate_node_properties(gds: GraphDataScience) -> None:
+    G, _ = gds.graph.project(GRAPH_NAME, {"Node": {"properties": ["x", "y"]}}, "*")
+
+    with pytest.raises(ValueError, match="Duplicate values in db_node_properties."):
+        gds.graph.nodeProperty.stream(
+            G, "x", db_node_properties=["z", "z", "name"], concurrency=2
+        )
+
+# @pytest.mark.enterprise
 @pytest.mark.compatible_with(min_inclusive=ServerVersion(2, 2, 0))
 def test_graph_nodeProperties_stream_with_arrow(gds: GraphDataScience) -> None:
     G, _ = gds.graph.project(GRAPH_NAME, {"Node": {"properties": ["x", "y"]}}, "*")
 
-    result = gds.graph.nodeProperties.stream(G, ["x", "y"], db_node_properties=["z", "name"], concurrency=2)
+    # also check duplicates will be filtered out
+    result = gds.graph.nodeProperties.stream(G, ["x", "x", "y"], db_node_properties=["z", "z", "name"], concurrency=2)
 
     assert list(result.keys()) == ["nodeId", "nodeProperty", "propertyValue"]
+    assert result.shape == (G.node_count() * 4, 3) # 4 properties
 
     x_values = result[result.nodeProperty == "x"]
     assert {e for e in x_values["propertyValue"]} == {1, 2, 3}
@@ -510,7 +520,7 @@ def test_graph_nodeProperties_stream_with_arrow_separate_property_columns(gds: G
     result = gds.graph.nodeProperties.stream(
         G, ["x", "y"], db_node_properties=["z", "name"], separate_property_columns=True, concurrency=2
     )
-    assert list(result.keys()) == ["nodeId", "x", "y", "z", "name"]
+    assert set(result.keys()) == {"nodeId", "x", "y", "z", "name"}
     assert {e for e in result["x"]} == {1, 2, 3}
     assert {e for e in result["y"]} == {2, 3, 4}
     assert len(result["z"]) == 3
@@ -549,10 +559,11 @@ def test_graph_nodeProperties_stream_without_arrow(gds_without_arrow: GraphDataS
     G, _ = gds_without_arrow.graph.project(GRAPH_NAME, {"Node": {"properties": ["x", "y"]}}, "*")
 
     result = gds_without_arrow.graph.nodeProperties.stream(
-        G, ["x", "y"], db_node_properties=["z", "name"], concurrency=2
+        G, ["x", "x", "y"], db_node_properties=["z", "z", "name"], concurrency=2
     )
 
-    assert {"nodeId", "nodeProperty", "propertyValue"}.issubset(set(result.keys()))
+    assert {"nodeId", "nodeProperty", "propertyValue", "nodeLabels"} == set(result.keys()) # empty nodeLabels column even if listNodeLabels = False
+    assert result.shape == (G.node_count() * 4, 3 + 1)  # 4 properties
 
     x_values = result[result.nodeProperty == "x"]
     assert {e for e in x_values["propertyValue"]} == {1, 2, 3}
@@ -568,6 +579,18 @@ def test_graph_nodeProperties_stream_without_arrow(gds_without_arrow: GraphDataS
     name_values = result[result.nodeProperty == "name"]
     assert {e for e in name_values["propertyValue"]} == {"nodeA", "nodeB", "nodeC"}
 
+def test_graph_nodeProperties_fail_on_duplicate_node_properties(gds: GraphDataScience) -> None:
+    G, _ = gds.graph.project(GRAPH_NAME, {"Node": {"properties": ["x", "y"]}}, "*")
+
+    with pytest.raises(ValueError, match="Duplicate values in node_properties"):
+        gds.graph.nodeProperties.stream(
+            G, ["x", "x", "y"], db_node_properties=["z", "name"], concurrency=2
+        )
+
+    with pytest.raises(ValueError, match="Duplicate values in db_node_properties."):
+        gds.graph.nodeProperties.stream(
+            G, ["x", "y"], db_node_properties=["z", "z", "name"], concurrency=2
+        )
 
 def test_graph_streamNodeProperties_without_arrow_separate_property_columns(
     gds_without_arrow: GraphDataScience,
@@ -598,7 +621,7 @@ def test_graph_nodeProperties_stream_without_arrow_separate_property_columns(
         G, ["x", "y"], db_node_properties=["z", "name"], separate_property_columns=True, concurrency=2
     )
 
-    assert list(result.keys()) == ["nodeId", "x", "y", "z", "name"]
+    assert set(result.keys()) == {"nodeId", "x", "y", "z", "name"}
 
     assert {e for e in result["x"]} == {1, 2, 3}
     assert {e for e in result["y"]} == {2, 3, 4}